# [Refactor] improve performance for detecting class components