--- a/lib/util/Components.js
+++ b/lib/util/Components.js
@@ -259,20 +259,29 @@ function componentRule(rule, context) {
   const utils = {
 
     /**
-     * Check if the node is a React ES5 component
+     * Check if an ObjectExpression is a React ES5 component
      *
      * @param {ASTNode} node The AST node being checked.
      * @returns {Boolean} True if the node is a React ES5 component, false if not
      */
     isES5Component(node) {
-      if (!node.parent) {
+      if (!node.parent || !node.parent.callee) {
         return false;
       }
-      return new RegExp(`^(${pragma}\\.)?${createClass}$`).test(sourceCode.getText(node.parent.callee));
+      const callee = node.parent.callee;
+      // React.createClass({})
+      if (callee.type === 'MemberExpression') {
+        return callee.object.name === pragma && callee.property.name === createClass;
+      }
+      // createClass({})
+      if (callee.type === 'Identifier') {
+        return callee.name === createClass;
+      }
+      return false;
     },
 
     /**
-     * Check if the node is a React ES6 component
+     * Check if a class is a React ES6 component
      *
      * @param {ASTNode} node The AST node being checked.
      * @returns {Boolean} True if the node is a React ES6 component, false if not
@@ -285,7 +294,14 @@ function componentRule(rule, context) {
       if (!node.superClass) {
         return false;
       }
-      return new RegExp(`^(${pragma}\\.)?(Pure)?Component$`).test(sourceCode.getText(node.superClass));
+      if (node.superClass.type === 'MemberExpression') {
+        return node.superClass.object.name === pragma
+          && /^(Pure)?Component$/.test(node.superClass.property.name);
+      }
+      if (node.superClass.type === 'Identifier') {
+        return /^(Pure)?Component$/.test(node.superClass.name);
+      }
+      return false;
     },
 
     /**