Benchmark on the [eslint repo](https://github.com/eslint/eslint) with `extends: plugin:react/recommended` and `time npx eslint .` (real time):

|  | before | after |
| --- | -- | -- |
| run 1 | 8.0s | 7.3s |
| run 2 | 7.8s | 7.4s |
| run 3 | 8.0s | 7.3s |

Benchmark with `TIMING=10 npx eslint .`:

```
Before:

Rule                           | Time (ms) | Relative
:------------------------------|----------:|--------:
react/no-deprecated            |   762.467 |    21.0%
react/no-direct-mutation-state |   648.951 |    17.9%
react/display-name             |   635.527 |    17.5%
react/no-string-refs           |   531.934 |    14.7%
react/prop-types               |   387.259 |    10.7%
react/require-render-return    |   382.931 |    10.6%
react/no-children-prop         |   165.876 |     4.6%
react/jsx-no-comment-textnodes |    49.452 |     1.4%
react/no-unescaped-entities    |    19.529 |     0.5%
react/jsx-key                  |    14.197 |     0.4%

After:

Rule                           | Time (ms) | Relative
:------------------------------|----------:|--------:
react/no-deprecated            |   531.083 |    21.5%
react/no-direct-mutation-state |   466.098 |    18.8%
react/display-name             |   427.566 |    17.3%
react/no-string-refs           |   313.779 |    12.7%
react/require-render-return    |   251.905 |    10.2%
react/prop-types               |   245.300 |     9.9%
react/no-children-prop         |   141.999 |     5.7%
react/jsx-no-comment-textnodes |    39.487 |     1.6%
react/no-unescaped-entities    |    17.461 |     0.7%
react/jsx-key                  |    12.571 |     0.5%
```